mlops_stack_taxi_fares/requirements.txt
mlflow>=2.14.0
numpy>=1.23.0
pandas>=1.4.3
scikit-learn>=1.1.1
matplotlib>=3.5.2
Jinja2>=3.0.3
# pyspark is pre-installed in Databricks runtime, commented out to avoid conflicts
# pyspark~=3.3.0
pytz~=2022.2.1
pyarrow>=14.0.0
databricks-feature-engineering>=0.13.0
lightgbm>=3.0.0

# Dev tools
black>=23.0.0
pylint>=2.17.0
mypy>=1.5.0
types-pytz>=2023.3.0.0